+++ src/network/model/packet-tag-list.h
@@ -228,8 +228,7 @@ class PacketTagList
      *
      * @param [in] buffer The buffer to read from.
      * @param [in] size The number of bytes to deserialize.
-     *
-     * @returns zero if complete tag list is not deserialized
+     * @return number of deserialized bytes on success, 0 on failure
      */
     uint32_t Deserialize(const uint32_t* buffer, uint32_t size);